Move EitherTransport into either.rs
And prettify imports of ipfs-private example
